Conflicts:
	src/renderer.cpp

+++ src/cube.cpp
@@ -1,21 +1,17 @@
 #include "cube.hpp"
 
-Cube::Cube(const float x, const float y, const float z, const float size)
+Cube::Cube(const double x, const double y, const double z, const double size)
 {   
     /* points specified from lower left corner, anti-clockwise */
     /* everything is on plane with z = 0.0f */
-    points[0] = {x, y, 0.0f, 1.0f};
-    points[1] = {x + size, y, 0.0f, 1.0f};
-    points[2] = {x + size, y + size, 0.0f, 1.0f};
-    points[3] = {x, y + size, 0.0f, 1.0f};
+    points[0] = {x, y, 0.0, 1.0};
+    points[1] = {x + size, y, 0.0, 1.0};
+    points[2] = {x + size, y + size, 0.0, 1.0};
+    points[3] = {x, y + size, 0.0, 1.0};
 
-    velocity = {0.0f, 0.0f, 0.0f, 0.0f};
+    velocity = {0.0, 0.0, 0.0, 0.0};
 } 
 
 Cube::~Cube()
 {
 }
-
-
-
-
