+++ src/cube.hpp
@@ -3,7 +3,7 @@
 
 class Cube {
 public:
-    Cube(const float x, const float y, const float z, const float size);
+    Cube(const double x, const double y, const double z, const double size);
     ~Cube();
 
     double points[4][4];
